refactoring build scripts
